add contact and change forms and bot
