+++ bundle/bots/listener/request/bot.yaml
@@ -1,3 +1,3 @@
-name: demorequest
+name: request
 type: LISTENER
 dialect: TYPESCRIPT
