+++ bundle/configvalues/email_template_contact_to_sales.yaml
@@ -1,0 +1,4 @@
+name: email_template_contact_to_sales
+public: true
+store: platform
+managedBy: site
