+++ bundle/configvalues/email_template_contact_to_user.yaml
@@ -1,0 +1,4 @@
+name: email_template_contact_to_user
+public: true
+store: platform
+managedBy: site
