+++ bundle/fields/uesio/crm/lead/source.yaml
@@ -1,0 +1,3 @@
+name: source
+type: TEXT
+label: Source
